backend/requirements.txt
# FairLens — Unified Backend Dependencies
# Install with: pip install -r requirements.txt

# ── Web framework (P1)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.6
python-dotenv>=1.0.1
pydantic>=2.7.1

# ── Data tools (P1 + P2)
pandas>=2.2.2
numpy>=1.26.4
joblib
scikit-learn>=1.5.0

# ── Fairness libraries (P2)
aif360>=0.6.1
fairlearn>=0.10.0

# ── SHAP (P2)
shap>=0.45.1

# ── PDF generation (P3)
weasyprint>=62.3
Jinja2>=3.1.3

# ── GCP (required when USE_LOCAL_STORAGE=false)
google-cloud-storage>=2.17.0
google-cloud-aiplatform>=1.53.0
vertexai>=1.53.0

# ── PDF fallback (used in pdf_generator.py when WeasyPrint unavailable)
reportlab>=4.0.0

# ── ONNX (optional — only needed for .onnx model inference)
# onnxruntime
